Adding separated button class. Initial temp data was added to payload.
