+++ src/device/esp8266_arduino/Button.h
@@ -1,0 +1,12 @@
+#ifndef ARDUINO_BUTTON_H
+#define ARDUINO_BUTTON_H
+
+#include <Arduino.h>
+
+/* Button activation in seconds. */
+#define BUTTON_ACT_SEC  3
+
+void ButtonInit();
+void ButtonLoop();
+
+#endif /* ARDUINO_BUTTON_H */
