+++ src/device/esp8266_arduino/Hardware.cpp
@@ -2,7 +2,9 @@
 
 #include "Hardware.h"
 #include "FileOrganizer.h"
+#include "PinMap.h"
 #include "TempSensor.h"
+#include "Button.h"
 #include "Global.h"
 
 /* Library lacks empty constructor.
@@ -10,35 +12,14 @@
  */
 OneWire oneWire(pinMap[2]);
 
-/* Button variables. */
-short hw_delay = 10;
-int   btn_current = 0;
-long  btn_millis_held = 0;
-long  btn_secs_held = 0;
-long  btn_prev_secs_held = 0;
-byte  btn_previous = HIGH;
-short btn_debouncing = 50;
-short btn_debounce_threshold = 200;
-unsigned long btn_first_time = 0;
-
-void ButtonInit()
-{
-#ifdef USE_NODEMCU_BOARD
-  pinMode(pinMap[3], INPUT);
-  /* NOTE: To work with pull-ups. */
-  digitalWrite(pinMap[3], HIGH);
-#elif defined(USE_DRAGON_BOARD)
-  pinMode(pinMap[4], INPUT);
-#else
-  pinMode(pinMap[4], INPUT);
-#endif
-}
-
 void GPIOInit()
 {
 #ifdef USE_NODEMCU_BOARD
   pinMode(pinMap[0], OUTPUT);
   pinMode(pinMap[1], OUTPUT);
+  
+  HWSetGPIO(0, false);
+  HWSetGPIO(1, false);
   /* pinMap[2] will be used for one-wire DQ. */
 #elif defined(USE_DRAGON_BOARD)
   pinMode(pinMap[6], OUTPUT);
@@ -48,49 +29,15 @@ void GPIOInit()
   pinMode(pinMap[7], OUTPUT);
   /* pinMap[2] will be used for one-wire DQ. */
 #endif
-  delay(hw_delay);
+
+  delay(HW_DELAY);
 }
 
 void SerialInit()
 {
   Serial.begin(SERIAL_BAUD_RATE);
   Serial.setDebugOutput(true);
-  delay(hw_delay);
-}
-
-void ButtonLoop()
-{
-  btn_current = digitalRead(pinMap[3]);
-  if (btn_current == LOW && btn_previous == HIGH && (millis() - btn_first_time) > btn_debounce_threshold) 
-  {
-    btn_first_time = millis();
-  }
-
-  btn_millis_held = (millis() - btn_first_time);
-  btn_secs_held = btn_millis_held / 1000;
-
-  if (btn_millis_held > btn_debouncing)
-  {
-    if (btn_current == LOW && btn_secs_held > btn_prev_secs_held)
-    {
-      Serial.printf("One second event!\n");
-    }
-
-    if (btn_current == HIGH && btn_previous == LOW) 
-    {
-      if (btn_secs_held <= 0)
-      {
-        Serial.printf("Pressed for less then 1 seconds!\n");
-      }
-      if (btn_secs_held >= BUTTON_ACT_SEC)
-      {
-        FSDeleteFile(WIFI_CONFIG_FILE);
-        ESP.restart();
-      }
-    }
-  }
-  btn_previous = btn_current;
-  btn_prev_secs_held = btn_secs_held;
+  delay(HW_DELAY);
 }
 
 void HWRestart()
@@ -101,14 +48,22 @@ void HWRestart()
 void HWSetGPIO(uint8_t pin, bool val)
 {
   /* Implement GPIO map for different devices. */
+#ifdef USE_LOW_LEVEL_TRIGGER
+  digitalWrite(pinMap[pin], !val);
+#else
   digitalWrite(pinMap[pin], val);
-  delay(hw_delay);
+#endif
+  delay(HW_DELAY);
 }
 
 bool HWGetGPIO(uint8_t pin)
 {
+#ifdef USE_LOW_LEVEL_TRIGGER
+  return !digitalRead(pinMap[pin]);
+#else
   /* Implement GPIO map for different devices. */
   return digitalRead(pinMap[pin]);
+#endif
 }
 
 void HWGetDeviceId(char *uid)
@@ -122,11 +77,15 @@ void HWInit()
   SerialInit();
   FSInit();
   ButtonInit();
+#ifdef USE_DS18B20_TEMP_SENSOR
   TEMPInit(&oneWire);
+#endif
 }
 
 void HWLoop()
 {
   ButtonLoop();
+#ifdef USE_DS18B20_TEMP_SENSOR
   TEMPLoop();
+#endif
 }
